Create a Task to configure devs namespaces
Developers will need to configure the private information relating
to their namespace (e.g. repository token). The goal of this
Task is to streamline the process to make it as painless as
possible.

The command to execute by the devs is given as an output the PEs when
installing the chart.

* Add a new Task that generates a secret with the required data
* Add the PipelineRun template that devs have to instantiate
* Output the command to be run by devs as part of the installing
* Add test to verify that the Task executes successfully
*