Improve DH deployment process
This commit should fix the DB migration lock error that sometimes
occur when deploying DH. The change to the Deployment resource
was triggering the creation of a new pod that could trigger a
race condition with the existing pod. The change to the resource
is now delayed until after the UI has reach a healthy state.