Fix security issue in Pipelines As Code
An attacker could submit a harmless PR, get the required 'ok-to-test'
to have the tekton pipelines executed, and then modify the PR to change
the content of the test pipelines.
Since the unreviewed pipeline would run by default, the attacker could
use it to gain access to data or resources they should not have access
to.